feat(web-page): split index.ts into modules, use webpack
affects: @practice/web-page

ISSUES CLOSED: #54
