+++ practice/web-page/.eslintrc
@@ -4,6 +4,12 @@
     "no-use-before-define": 0,
     "no-undef": 0,
     "no-return-assign": 0,
-    "no-param-reassign": 0
+    "no-param-reassign": 0,
+    "import/extensions": 0
+  },
+  "settings": {
+    "import/resolver": {
+      "typescript": "./src/**/*"
+    }
   }
 }
