+++ practice/web-page/package.json
@@ -5,7 +5,15 @@
   "main": "dist/index.js",
   "source": "src/index.ts",
   "scripts": {
-    "build": "tsc",
+    "build": "webpack",
+    "compile": "tsc",
     "lint": "eslint"
-  }
+  },
+    "devDependencies": {
+      "eslint-import-resolver-typescript": "^2.4.0",
+      "ts-loader": "^8.0.17",
+      "ts-node": "^9.1.1",
+      "webpack": "^5.22.0",
+      "webpack-cli": "^4.5.0"
+    }
 }
