+++ practice/web-page/public/index.html
@@ -37,7 +37,7 @@
         <div class="container">
             <div class="tabcontainer">
                 <div class="tabcontent">
-                    <img src='../src/assets/img/tabs/vegy.jpg' alt="vegy">
+                    <img src="../src/assets/img/tabs/vegy.jpg" alt="vegy">
                     <div class="tabcontent__description">
                         Меню "Фитнес" - это новый подход к приготовлению блюд: больше свежих овощей и фруктов. Для людей, которые интересуются спортом; активных и здоровых. Это абсолютно новый продукт с оптимальной ценой и высоким качеством!
                     </div>
@@ -271,7 +271,7 @@
         </div>
     </div>
 
-<script src='../dist/index.js'></script>
+<script src='../dist/bundle.js'></script>
 
 </body>
 </html>
