+++ practice/web-page/src/modules/menuCards.ts
@@ -1,0 +1,78 @@
+import { getResources } from '../services/services'
+
+function menuCards() {
+  class MenuCard {
+    imgSrc: string
+
+    alt: string
+
+    title: string
+
+    description: string
+
+    price: number
+
+    parentSelector: HTMLElement
+
+    classes: string[]
+
+    currentExchangeRate: number
+
+    element: string
+
+    constructor(
+      imgSrc: string,
+      alt: string,
+      title: string,
+      description: string,
+      price: number,
+      parentSelector: string,
+      ...classes: string[]
+    ) {
+      this.imgSrc = imgSrc
+      this.alt = alt
+      this.title = title
+      this.description = description
+      this.price = price
+      this.parentSelector = document.querySelector(parentSelector)
+      this.classes = classes
+      this.currentExchangeRate = 75
+      this.convertCurrency()
+      this.element = 'menu__item'
+    }
+
+    convertCurrency() {
+      this.price *= this.currentExchangeRate
+    }
+
+    render() {
+      const element: HTMLElement = document.createElement('div')
+
+      if (this.classes.length === 0) {
+        element.classList.add(this.element)
+      } else {
+        this.classes.forEach((className) => element.classList.add(className))
+      }
+
+      element.innerHTML = `
+          <img src=${this.imgSrc} alt=${this.alt}>
+          <h3 class="menu__item-subtitle">${this.title}</h3>
+          <div class="menu__item-description">${this.description}</div>
+          <div class="menu__item-divider"></div>
+          <div class="menu__item-price">
+            <div class="menu__item-cost">Цена:</div>
+            <div class="menu__item-total"><span>${this.price}</span> руб/день</div>
+          </div>
+      `
+      this.parentSelector.append(element)
+    }
+  }
+
+  getResources('https://6027afc0dd4afd001754a9b0.mockapi.io/api/menu/').then((data) => {
+    data.forEach(({ imgSrc, alt, title, description, price }) => {
+      new MenuCard(imgSrc, alt, title, description, price, '.menu .container').render()
+    })
+  })
+}
+
+export default menuCards
